refactor: 기존 화면 composable의 이름을 CalendarMainScreen으로 변경
